pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ainotewriter"
version = "0.1.0"
description = "X Community Notes AI Writer with CLI, GitHub Actions, and desktop app"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "requests>=2.32.3",
  "requests-oauthlib>=2.0.0",
  "python-dotenv>=1.0.1",
  "claude-code-sdk>=0.0.14"
]

[project.scripts]
ai-note-writer = "ainotewriter.cli:main"
ai-note-writer-gui = "ainotewriter.gui:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
